+++ examples/neural_compressor/optical-character-recognition/requirements.txt
@@ -1,0 +1,3 @@
+datasets >= 1.8.0
+pillow
+torch >= 1.9.0
